ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain
google-genai
google-generativeai
chromadb
sentence-transformers
websockets
firebase-admin

# OpenAI Realtime API (alternative to Gemini)
openai>=1.12.0
